LAB7_OPENCV/requirements.txt
absl-py==1.2.0
astunparse==1.6.3
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
PythonQwt==0.9.0
pytz==2022.5
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
six==1.16.0
termcolor==2.0.1
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
absl-py==1.2.0
asttokens==2.2.1
astunparse==1.6.3
autopep8==2.0.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
h5py==3.7.0
heartpy==1.2.7
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
Pygments==2.14.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PythonQwt==0.9.0
pytz==2022.5
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.8.1
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
absl-py==1.2.0
asttokens==2.2.1
astunparse==1.6.3
autopep8==2.0.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
h5py==3.7.0
heartpy==1.2.7
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
Pygments==2.14.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PythonQwt==0.9.0
pytz==2022.5
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.8.1
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
absl-py==1.2.0
asttokens==2.2.1
astunparse==1.6.3
autopep8==2.0.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
executing==1.2.0
flatbuffers==22.9.24
fonttools==4.37.4
gast==0.4.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.49.1
h5py==3.7.0
heartpy==1.2.7
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
Pygments==2.14.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PythonQwt==0.9.0
pytz==2022.5
QtPy==2.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
six==1.16.0
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tesseract-ocr==0.0.1
threadpoolctl==3.1.0
tomli==2.0.1
traitlets==5.8.1
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
